Revert "changed trend calc float to int (precision not needed)"
This reverts commit e85ccf97277fbfbead67483e1d11ed9e2d425f69.

+++ WiFiWeatherDisplayEsp32.ino
@@ -214,7 +214,7 @@ void displayCovidData() {
   displayTrend(getTrend(readFileString(PATH_RVALUE_PREV).toFloat(), readFileString(PATH_RVALUE).toFloat()), 2000);
 }
 
-uint8_t getTrend(int val1, int val2) {
+uint8_t getTrend(float val1, float val2) {
   // TODO: add margin
   // 0 = error; 1 = equal; 2 = falling; 3 = rising
   if(val1 == val2) {
